Move theme to material2 module
