+++ core/ui/compose/testing/build.gradle.kts
@@ -11,7 +11,7 @@ dependencies {
     api(libs.turbine)
     api(libs.assertk)
 
-    implementation(projects.core.ui.compose.theme)
+    implementation(projects.core.ui.compose.theme.material2)
     implementation(libs.androidx.compose.material)
 
     implementation(libs.bundles.shared.jvm.test.compose)
