+++ core/ui/compose/theme/material2/src/main/java/app/k9mail/core/ui/compose/theme/material2/MainTheme.kt
@@ -1,4 +1,4 @@
-package app.k9mail.core.ui.compose.theme
+package app.k9mail.core.ui.compose.theme.material2
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
@@ -6,10 +6,10 @@ import androidx.compose.material.Typography
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.ReadOnlyComposable
-import app.k9mail.core.ui.compose.theme.color.Colors
-import app.k9mail.core.ui.compose.theme.color.LocalColors
-import app.k9mail.core.ui.compose.theme.color.SetSystemBarsColor
-import app.k9mail.core.ui.compose.theme.color.toMaterialColors
+import app.k9mail.core.ui.compose.theme.material2.color.Colors
+import app.k9mail.core.ui.compose.theme.material2.color.LocalColors
+import app.k9mail.core.ui.compose.theme.material2.color.SetSystemBarsColor
+import app.k9mail.core.ui.compose.theme.material2.color.toMaterialColors
 
 @Composable
 fun MainTheme(
