+++ core/ui/compose/theme/material2/src/main/java/app/k9mail/core/ui/compose/theme/material2/Sizes.kt
@@ -1,4 +1,4 @@
-package app.k9mail.core.ui.compose.theme
+package app.k9mail.core.ui.compose.theme.material2
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.staticCompositionLocalOf
