+++ core/ui/compose/theme/material2/src/main/java/app/k9mail/core/ui/compose/theme/material2/color/SystemBarsColor.kt
@@ -1,4 +1,4 @@
-package app.k9mail.core.ui.compose.theme.color
+package app.k9mail.core.ui.compose.theme.material2.color
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
