+++ core/ui/compose/theme/material2/src/main/res/drawable/core_ui_theme_k9_logo.xml
