+++ core/ui/compose/theme/material2/src/main/res/drawable/core_ui_theme_thunderbird_logo.xml
